Pagination fully implemented & Code cleaned
