+++ src/main/java/com/breadsb/notesapp/NotesappApplication.java
@@ -7,7 +7,6 @@ import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfi
 import org.springframework.context.annotation.PropertySource;
 
 @SpringBootApplication
-@PropertySource("classpath:application-local.properties")
 public class NotesappApplication {
 
 	public static void main(String[] args) {
