+++ src/main/java/com/breadsb/notesapp/controllers/NoteController.java
@@ -26,8 +26,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/notes/")
-//@RequiredArgsConstructor
+@RequestMapping("/api/notes")
 @CrossOrigin("*")
 public class NoteController {
 
@@ -35,14 +34,13 @@ public class NoteController {
     private final NoteService service;
     @Autowired
     private final NotePaginatedService paginatedService;
-    private final String API_PATH = "${pageContext.request.requestURL}";
     Bucket bucket;
 
     public NoteController(NoteService service, NotePaginatedService paginatedService) {
         this.service = service;
         this.paginatedService = paginatedService;
         bucket = Bucket.builder()
-                .addLimit(limit -> limit.capacity(6).refillGreedy(6,Duration.ofMinutes(1)))
+                .addLimit(limit -> limit.capacity(10).refillGreedy(6,Duration.ofMinutes(1)))
                 .build();
     }
 
@@ -57,7 +55,7 @@ public class NoteController {
             @ApiResponse(responseCode = "400", description = "Invalid id supplied", content = @Content),
             @ApiResponse(responseCode = "404", description = "Book not found", content = @Content)
     })
-    @GetMapping("{id}")
+    @GetMapping("/{id}")
     public ResponseEntity<Note> get(@Parameter(description = "Id of a book to search") @PathVariable Long id) {
         if (bucket.tryConsume(1)) {
             return ResponseEntity.ok(service.getNoteById(id));
@@ -75,8 +73,6 @@ public class NoteController {
     public ResponseEntity<Void> post(@io.swagger.v3.oas.annotations.parameters.RequestBody(description = "Note JSON object to be transfered") @Valid @RequestBody Note note) {
         if (bucket.tryConsume(1)) {
             service.createNote(note);
-//            String uri = API_PATH + note.getId();
-//            return ResponseEntity.created(URI.create(uri)).build();
             return ResponseEntity.ok().build();
         }
         return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
@@ -106,13 +102,11 @@ public class NoteController {
             @ApiResponse(responseCode = "400", description = "Invalid id format", content = @Content),
             @ApiResponse(responseCode = "415", description = "No object has been sent (null value)", content = @Content)
     })
-    @PutMapping(value = "{id}", consumes = MediaType.APPLICATION_JSON_VALUE)
+    @PutMapping(value = "/{id}", consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Note> put(@Parameter(description = "Id of a Note to update") @PathVariable Long id,
                                     @io.swagger.v3.oas.annotations.parameters.RequestBody(description = "Note JSON object") @Valid @RequestBody Note note) {
         if (bucket.tryConsume(1)) {
             service.updateNote(id, note);
-//            String uri = API_PATH + id;
-//            return ResponseEntity.created(URI.create(uri)).build();
             return ResponseEntity.ok().build();
         }
         return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
@@ -124,7 +118,7 @@ public class NoteController {
             @ApiResponse(responseCode = "404", description = "Note not found", content = @Content),
             @ApiResponse(responseCode = "400", description = "Invalid id format", content = @Content)
     })
-    @DeleteMapping("{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@Parameter(description = "Id of a Note to delete") @PathVariable Long id) {
         service.deleteNoteById(id);
         return ResponseEntity.ok().build();
@@ -136,7 +130,7 @@ public class NoteController {
             @ApiResponse(responseCode = "404", description = "Notes not found", content = @Content),
             @ApiResponse(responseCode = "415", description = "Null DateTime parameter", content = @Content)
     })
-    @GetMapping(value = "by-date/")
+    @GetMapping(value = "/by-date")
     public ResponseEntity<List<Note>> getNotesByDate(
             @Parameter(description = "Time and date when notes were created")
             @RequestParam("timestamp")
@@ -147,7 +141,7 @@ public class NoteController {
         return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
     }
 
-    @GetMapping("page")
+    @GetMapping("/page")
     public ResponseEntity<Page<Note>> getPaginatedNotes(Pageable pageable) {
         if (bucket.tryConsume(1)) {
             Page<Note> notes = paginatedService.getNotesPage(pageable);
@@ -156,7 +150,7 @@ public class NoteController {
         return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
     }
 
-    @GetMapping("page/count")
+    @GetMapping("/page/count")
     public ResponseEntity<Integer> getNumberOfPages(@RequestParam(defaultValue = "10") int size) {
         if (bucket.tryConsume(1)) {
             int number = paginatedService.getNumberOfPages(size);
