+++ src/main/resources/application.properties
@@ -1,17 +1,18 @@
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
-springdoc.packagesToScan=com.breadsb.controllers
+springdoc.packagesToScan=com.breadsb.notesapp.controllers
 
-#spring.config.import=sm://
+spring.config.import=sm://
 spring.cloud.gcp.sql.database-name=notes
-#spring.cloud.gcp.sql.instance-connection-name=${sm://projects/130607545475/secrets/gcloud_database_instance}
+spring.cloud.gcp.sql.instance-connection-name=${sm://projects/130607545475/secrets/gcloud_database_instance}
 spring.datasource.continue-on-error=true
 spring.datasource.initialization-mode=always
 spring.datasource.username=postgres
-#spring.datasource.password=${sm://projects/130607545475/secrets/gcloud_notesdb_pass}
+spring.datasource.password=${sm://projects/130607545475/secrets/gcloud_notesdb_pass}
 spring.datasource.hikari.maximum-pool-size=10
 spring.datasource.hikari.minimum-idle=1
 
 spring.mvc.favicon.enabled=false
 #server.port=8443
+security.basic.enabled=false
