+++ src/main/resources/templates/app/view-messages.html
@@ -30,9 +30,6 @@
 
     <section>
         <div class="buttons">
-            <form action="https://noteapp-399113.appspot.com/api/notes/" method="get" id="getAllMessagesForm">
-                <button name="getAllMessages" data-button-get-all-messages type="button" class="button">Fetch messages</button>
-            </form>
             <button id="createMessageButton" class="button">Create message</button>
         </div>
         <div id="paginationMenu"></div>
